+++ composenav/src/main/java/com/adamkobus/compose/navigation/action/PopAction.kt
@@ -16,8 +16,8 @@ class PopAction(
     /**
      * Calls [NavHostController.popBackStack]
      */
-    override fun navigate(controller: NavHostController) {
-        controller.popBackStack()
+    override fun navigate(controller: NavHostController): Boolean {
+        return controller.popBackStack()
     }
 
     /**