+++ demo-ui-dogs-browser/src/main/java/com/adamkobus/compose/navigation/demo/ui/dogdetails/DogDetailsScreenInteractions.kt
@@ -1,5 +1,6 @@
 package com.adamkobus.compose.navigation.demo.ui.dogdetails
 
 data class DogDetailsScreenInteractions(
-    val onOpenDialogClicked: () -> Unit = {}
+    val onOpenDialogClicked: () -> Unit = {},
+    val onOpenGalleryClicked: () -> Unit = {}
 )