+++ demo-ui-dogs-browser/src/main/java/com/adamkobus/compose/navigation/demo/ui/doggallery/DogGalleryScreenInteractions.kt
@@ -1,0 +1,5 @@
+package com.adamkobus.compose.navigation.demo.ui.doggallery
+
+data class DogGalleryScreenInteractions(
+    val onBackToListClicked: () -> Unit
+)